Add support for revision filering
Users now can specify filters based on revisions, it can be single
revision in the form of either NNN or rNNN where NNN is the revision
number or a range in the form of NNN-MMM or rNNN-rMMM. For the range
filter both start and end revisions are included in the filter.

Multiple revision filters can be specified by space-separating them.
In this case they are compined using OR operator

Closes #1